Tutor - Merged the add & edit components
